+++ helpers/index.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2017, Shopgate, Inc. All rights reserved.
+ * Copyright (c) 2017-present, Shopgate, Inc. All rights reserved.
  *
  * This source code is licensed under the Apache 2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -9,7 +9,6 @@ import get from 'lodash/get';
 import find from 'lodash/find';
 import core from '@shopgate/tracking-core/core/Core';
 import { logger } from '@shopgate/pwa-core/helpers';
-import event from '@shopgate/pwa-core/classes/Event';
 
 /**
  * Converts a price to a formatted string.
@@ -143,21 +142,6 @@ export const formatPurchaseData = (passedOrder) => {
 };
 
 /**
- * Flag to enable/disable the tracking.
- * @type {boolean}
- */
-let trackingDisabled = false;
-
-// Disable the tracking if the webview is not in the foreground anymore
-event.addCallback('viewWillDisappear', () => {
-  trackingDisabled = true;
-});
-// Enable the tracking if the webview enters the foreground
-event.addCallback('viewWillAppear', () => {
-  trackingDisabled = false;
-});
-
-/**
  * Helper to pass the redux state to the tracking core
  * @param {string} eventName The name of the event.
  * @param {Object} data The tracking data of the event.
@@ -167,11 +151,6 @@ event.addCallback('viewWillAppear', () => {
 export const track = (eventName, data, state) => {
   if (typeof core.track[eventName] !== 'function') {
     logger.warn('Unknown tracking event:', eventName);
-    return false;
-  }
-
-  if (trackingDisabled) {
-    logger.warn('Tracking disabled');
     return false;
   }